LPS-57828 Move SerializableObjectWrapper from portal-cache-ehcache to portal-cache
