+++ modules/portal-cache/portal-cache-ehcache/src/com/liferay/portal/cache/ehcache/internal/SerializableEhcachePortalCache.java
@@ -14,7 +14,7 @@
 
 package com.liferay.portal.cache.ehcache.internal;
 
-import com.liferay.portal.cache.ehcache.SerializableObjectWrapper;
+import com.liferay.portal.cache.io.SerializableObjectWrapper;
 import com.liferay.portal.kernel.cache.PortalCacheManager;
 
 import java.io.Serializable;
