+++ modules/portal-cache/portal-cache-ehcache/src/com/liferay/portal/cache/ehcache/internal/event/EhcachePortalCacheListenerAdapter.java
@@ -14,8 +14,8 @@
 
 package com.liferay.portal.cache.ehcache.internal.event;
 
-import com.liferay.portal.cache.ehcache.SerializableObjectWrapper;
 import com.liferay.portal.cache.ehcache.internal.EhcacheUnwrapUtil;
+import com.liferay.portal.cache.io.SerializableObjectWrapper;
 import com.liferay.portal.kernel.cache.PortalCache;
 import com.liferay.portal.kernel.cache.PortalCacheException;
 import com.liferay.portal.kernel.cache.PortalCacheListener;
